Refactor and decoupling for subcomponents
In preparation of unit testing `com.simperium.Simperium` it became necessary to
be able to inject fake components for the networking and persistence stacks.

There are now four well defined components for a Simperium client which is
presented in the `com.simperium.client.ClientFactory` class.

Still a work in progress but the more defined these components and the more
decoupled the easier it will be to start supporting alternate platforms.