JSONDiff uses it's own equals method to compare arrays and objects
Two `JSONObject`'s with same key-value pairs but in a different order are still equal. And two arrays containing these respective objects are equal as well.

For now we are using the string representations of these objects for equality comparison since they sort the keys alpha-numerically but it is potentially a performance issue.

Fixes #57