Switching change procesor to perform one change per execution
For each remote change or local change to process it will only do one unit of work.

1. Check remote changes, if there is one, process it and schedule another run
2. If no remote chagnes, check local changes to send out, if there is one, send it out
3. If there are no remote changes and no local changes it can stop doing work

When a new remote change or local change is queued, the processor gets added to the executor and will start working again.

Retrying changes are handled outside of this work and will fire at the scheduled timer interval.

Fixes #53