MockExecutor provides Immediate and Playable executor services
Immediate runs each runnable as it receives it. This effectively makes the Executor single threaded within the context it is being used. Useful for tests that are not concerned with concurrency.

Playable queues up each Runnable and then executes them in turn when Playable.play is called. Any Runnable's added to the executor during play will also be executed. Useful for testing concurrency since the test can control when the Runnables are executed.