After applying a remote change clean up pending changes
After a remote change is applied there can be a local change queued to go out.

If the incoming change removed the object, remove all pending changes. If the incoming change modified the object, recalculate the change to sync.

Tell the serializer to dequeue the pending change that will no longer be sent.