Add heartbeat listener interface for detecting server communication status
Previously we haven't offered any way to determine if the connection with the
Simperium server is up or not. With this change we're exposing a new listener
that a client application can use to be pinged whenever we get a heartbeat
signal from the server, indicating that a connection is active.