Bucket: Generalize listener notification pattern
In each bucket event listener notifier method (`notifyOnSaveListeners` and friends)
we have been repeating the general logic of iterating the listeners for that type
and then catching and logging exceptions thrown during the notification.

In this patch we are creating a generalized listener-iterator method `notifyListeners`
which is then used by each of the listener notifier methods so they can stick to
their specific behaviors.

While the code change count is almost zero and we've created one new method we have
moved the redundant bits about the act of iterating over the listeners into one place
where we aren't at risk of copy/paste errors. The line count would have decreased
were it not for the fact that we require three and a half lines just to setup the
`Consumer` function that performs the specific logic in each method.

`androidx.core.util.Consumer` was chosen for compatability since `java.util.Consumer`
was added in Android SDK version 24.