Name post-processing conditions at end of processRemoteChange()
At the end of `processRemoteChange()` we have been deciding whether or
not to purge remaining local changes from the queue if we receive a change
that removes the associated entity.

In this branch we've started also notifying listeners of updates to an
entity as sync events. Now, however, the logic at the end feels jumbled
and the two look related somehow.

In this commit I've given semantic names to the conditions used to
trigger each side effect in hopes that it clears up the relationship
between each one and the change, and between each other.