Change sendRequest to handle non-json results in case of error
The previous callback to handle the response of requests was using the
JSON parser by default. The problem was that in responses where the code
wasn't 200, the response was not a JSON object. Thus, we changed the
callback to be a StringCallback and parse the JSON object in case the
response is 200. Also, a Throwable is sent in case of error with the
body as message of the Throwable.